Prevent synthetic case class members from marking other synthetic members of the same case class as used.
This fixes an issue where a case class with at least one field is always considered used. It happens because `def productElement`, a synthetic method, refers to `def _1`, another synthetic method, which refers to the case class itself.

The code was already setup so that a reference to a case class from a synthetic method would not cause the case class to be considered used -- it works by building up a `usedProxy` map where the keys are the symbols of the synthetic methods and the values are the symbols of the case classes. At the end of the world, if any of the symbols that appear in the keys of `usedProxy` have been used, then we consider the values sued as well.

However this didn't account for cases where the `usedProxy` keys (the synthetic methods) are used by other synthetic methods. Since this happens with `productElement` and `_1`, case classes were never reported unused.